+++ mycodo/outputs/pwm_pca9685.py
@@ -169,13 +169,6 @@ OUTPUT_INFORMATION = {
             'phrase': 'Invert the value that is saved to the measurement database'
         },
         {
-            'id': 'trigger_functions_startup',
-            'type': 'bool',
-            'default_value': False,
-            'name': lazy_gettext('Trigger Functions at Startup'),
-            'phrase': 'Whether to trigger functions when the output switches at startup'
-        },
-        {
             'id': 'amps',
             'type': 'float',
             'default_value': 0.0,