+++ mycodo/outputs/pwm_shell.py
@@ -124,13 +124,6 @@ OUTPUT_INFORMATION = {
             'phrase': 'Invert the value that is saved to the measurement database'
         },
         {
-            'id': 'trigger_functions_startup',
-            'type': 'bool',
-            'default_value': False,
-            'name': lazy_gettext('Trigger Functions at Startup'),
-            'phrase': lazy_gettext('Whether to trigger functions when the output switches at startup')
-        },
-        {
             'id': 'command_force',
             'type': 'bool',
             'default_value': False,